pyproject.toml
[tool.poetry]
name = "herald"
version = "0.1.0"
description = ""
authors = ["Pavel <hydrophonyx@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
plotly = "5.18.0"
attrs = "^23.2.0"
dvc = {extras = ["gdrive"], version = "^3.37.0"}
lightning = "2.1"
torch = "^2.1.2"
pytorch-tabnet = "^4.1.0"
pillow = "^10.2.0"
hydra-core = "^1.3.2"
skl2onnx = "^1.16.0"
onnxruntime = "^1.16.3"
mlflow = "^2.9.2"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["herald"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
